refactor ProductsImport extension to controller structure
