+++ hugashop/crm/Extensions/ProductsImport/settings.yaml
@@ -1,3 +1,3 @@
 name: "Импорт товаров на склад"
 description: "Загрузка CSV-файла с товарами и добавление их на выбранный склад"
-version: 1.2
+version: 1.3
